learning how to make a commit
To me,  making a commit is like adding a change into the branch.

In the whole Github Spectrum:
The repository is a tree. It is a tree that can hold coding projects and images and pretty much anything.

With branches, you are making a part of that main repository to be used for edits and remakes. Basically, with branches you can use them to edit files.

Imagine that the commits are the leaves. Because of this, you will be able to change up your branch.